File Pointers and Function Definitions
Rearranged some constants, created file names pointers and opened them. Started to get info from the first file (flights.txt) and call function with the line read from the file. Did error checking for functions used so far, also a bit of pseudocode for the processFlight() function, still need to work on parseLine() function before processFlight() can be done. There's a bunch of rough drafts commented out from when I started, kept them in just in case i can use parts of it or refer back to it. Also comments throughout as reminders or questions to self to answer later.