Display Function + Fixing Logical Errors
Finally figured out how to sort through the struct array and only print the source-destination pair with the lowest fare. The 2 biggest factors of change were: the for loop within the for loop where the index should be initialized to 0 and not the (previous for loops incrementing value) + 1 in order to always compare to the whole array to the current index; and the second being printing the flight data only if currentFlightIndex == indexLowestPrice.

The biggest logical error was within my while loops. I don't know how i didn't understand it before... But basically the condition was || when it should have been &&. It took a bit to realize that both conditions need to be true in order to keep processing data, or else it will go on forever, not stopping when the total lines to read hits 100.

Also added in my own error checking for spaces before and after strings that come from flights.txt. In my previous code, an airline.txt wouldn't open if there were spaces before the text, however i realized that in the example it does. So i went back and spent another hour on it LOL